try to seed join table
