+++ app/models/__init__.py
@@ -1,4 +1,4 @@
 from .db import db
-from .user import User
-from .servers import Server
+from .user import User, Server
+# from .servers import Server
 from .db import environment, SCHEMA
