+++ app/models/servers.py
@@ -1,35 +1,65 @@
-from .db import db, environment, SCHEMA, add_prefix_for_prod
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.schema import Column, ForeignKey, Table
+# from .db import db, environment, SCHEMA, add_prefix_for_prod
+# from sqlalchemy.ext.declarative import declarative_base
+# from sqlalchemy.schema import Column, ForeignKey, Table
+# from sqlalchemy.orm import relationship
+# from .db import add_prefix_for_prod
 
 
 
-Base = declarative_base()
+# Base = declarative_base()
 
-server_users = Table(
-    'server_users',
-    Base.metadata,
-    Column('serverId', ForeignKey('servers.id'), primary_key=True),
-    Column('userId', ForeignKey('users.id'), primary_key=True),
-)
+# server_users = Table(
+#     'server_users',
+#     Base.metadata,
+#     db.Column('serverId', db.Integer, ForeignKey(add_prefix_for_prod('servers.id')), primary_key=True),
+#     db.Column('userId', db.Integer, ForeignKey(add_prefix_for_prod('users.id')), primary_key=True),
+# )
 
-class Server(db.Model):
-    __tablename__ = 'servers'
+# server_admins = Table(
+#     'server_admins',
+#     Base.metadata,
+#     db.Column('serverId',db.Integer, ForeignKey(add_prefix_for_prod('servers.id')), primary_key=True),
+#     db.Column('userId', db.Integer, ForeignKey(add_prefix_for_prod('users.id')), primary_key=True),
+# )
 
-    if environment == "production":
-        __table_args__ = {'schema': SCHEMA}
+# class Server(db.Model):
+#     __tablename__ = 'servers'
 
-    id = db.Column(db.Integer, primary_key=True)
-    ownerId = db.Column(db.Integer, nullable=False)
-    name = db.Column(db.String(255), nullable=False)
-    preview_img = db.Column(db.String(255))
-    admin_Ids = db.Column(db.String(255))
+#     if environment == "production":
+#         __table_args__ = {'schema': SCHEMA}
 
-    def to_dict(self):
-        return {
-            'id': self.id,
-            'ownerId': self.ownerId,
-            'name': self.name,
-            'preview_img': self.preview_img,
-            'admin_Ids': self.admin_Ids
-        }
+#     id = db.Column(db.Integer, primary_key=True)
+#     ownerId = db.Column(db.Integer, nullable=False)
+#     name = db.Column(db.String(255), nullable=False)
+#     preview_img = db.Column(db.String(255))
+    
+#     users = relationship('User',
+#                         secondary = server_users,
+#                         back_populates = 'servers')
+    
+#     admins = relationship('User',
+#                         secondary=server_admins,
+#                         back_populates='servers')
+
+# # class Server_user(db.Model):
+# #     __tablename__ = 'server_users'
+    
+# #     serverId = db.Column(db.Integer, db.ForeignKey('servers.id'))
+# #     userId = db.Column(db.Integer, db.ForeignKey('users.id'))
+    
+    
+    
+# # class Server_admin(db.Model):
+# #     __tablename__ = 'server_admins'
+    
+# #     serverId = db.Column(db.Integer, db.ForeignKey('servers.id'))
+# #     userId = db.Column(db.Integer, db.ForeignKey('users.id'))
+
+#     def to_dict(self):
+#         return {
+#             'id': self.id,
+#             'ownerId': self.ownerId,
+#             'name': self.name,
+#             'preview_img': self.preview_img,
+            
+#         }
